+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 __New Features__
 * [#708](https://github.com/nghuuphuoc/bootstrapvalidator/issues/708), [#899](https://github.com/nghuuphuoc/bootstrapvalidator/issues/899): Add ```setLocale()``` and ```getLocale()``` methods to support multiple languages
 * [#718](https://github.com/nghuuphuoc/bootstrapvalidator/issues/718): Add ```validateContainer()``` method
+* [#744](https://github.com/nghuuphuoc/bootstrapvalidator/issues/744): Add ```transformer``` option, allowing to hook the value of field before validating
 * [#1131](https://github.com/nghuuphuoc/bootstrapvalidator/issues/1131): Support add-on
 * [#1140](https://github.com/nghuuphuoc/bootstrapvalidator/pull/1140): Add UAE [phone number](http://bootstrapvalidator.com/validators/phone/) validator, thanks to [@s-a-y](https://github.com/s-a-y)
 * [#1153](https://github.com/nghuuphuoc/bootstrapvalidator/pull/1153): Add EIN validator, thanks to [@joshuachestang](https://github.com/joshuachestang)